backend/requirements-django-simple.txt
django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-multipart==0.0.6
python-dotenv==1.0.0
